[node_pushserver] improved support for project configure api
